+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Travis config file locally and in template
 - CHANGELOG.md file to keep a changelog locally and in template
 
 ## [1.0.2] - 2019-01-31